requirements.txt
# Core dependencies
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio==3.4.3

# Data processing
pydantic==2.5.2
python-dotenv==1.0.0

# HTTP and API clients
requests==2.31.0
httpx==0.25.2

# Date and time handling
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0

# Environment and configuration
python-decouple==3.8

# Web framework (for health checks)
fastapi==0.104.1
uvicorn==0.24.0

# Additional utilities
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0